+++ src/ReducedModelControllers/src/CentroidalMPC.cpp
@@ -556,6 +556,8 @@ struct CentroidalMPC::Impl
 
         // resize the CoM Trajectory
         this->output.comTrajectory.resize(stateHorizon);
+        this->output.comVelocityTrajectory.resize(stateHorizon);
+        this->output.angularMomentumTrajectory.resize(stateHorizon);
 
         // In case of no warmstart the variables are:
         // - centroidalVariables = 7: external force + external torque + com current + dcom current
@@ -1071,6 +1073,8 @@ struct CentroidalMPC::Impl
         }
 
         concatenateOutput(this->optiVariables.com, "com");
+        concatenateOutput(this->optiVariables.dcom, "dcom");
+        concatenateOutput(this->optiVariables.angularMomentum, "angular_momentum");
 
         casadi::Dict toFunctionOptions, jitOptions;
         if (casadiVersionIsAtLeast360())
@@ -1263,6 +1267,21 @@ bool CentroidalMPC::advance()
         using namespace BipedalLocomotion::Conversions;
         m_pimpl->output.comTrajectory[i] = toEigen(*it).col(i);
     }
+
+    std::advance(it, 1);
+    for (int i = 0; i < m_pimpl->output.comVelocityTrajectory.size(); i++)
+    {
+        using namespace BipedalLocomotion::Conversions;
+        m_pimpl->output.comVelocityTrajectory[i] = toEigen(*it).col(i);
+    }
+
+    std::advance(it, 1);
+    for (int i = 0; i < m_pimpl->output.angularMomentumTrajectory.size(); i++)
+    {
+        using namespace BipedalLocomotion::Conversions;
+        m_pimpl->output.angularMomentumTrajectory[i] = toEigen(*it).col(i);
+    }
+
     // advance the time
     m_pimpl->currentTime += m_pimpl->optiSettings.samplingTime;